requirements.txt
django==1.11.4
django-debug-toolbar
django-extensions
django-bootstrap3==9.0.0
django-el-pagination==3.1.0
pilkit==2.0
ipython[notebook]
request
django-imagekit


